Update and cleanup actors
