+++ vm/actor/src/builtin/account/state.rs
@@ -1,0 +1,9 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use address::Address;
+
+/// AccountActorState includes the address for the actor
+pub struct AccountActorState {
+    pub address: Address,
+}
