+++ vm/actor/src/builtin/init/mod.rs
@@ -1,61 +1,14 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::FIRST_NON_SINGLETON_ADDR;
-use vm::{
-    ActorID, CodeID, ExitCode, MethodNum, Serialized, METHOD_CONSTRUCTOR, METHOD_PLACEHOLDER,
-};
+mod state;
 
+pub use self::state::InitActorState;
 use address::Address;
-use cid::Cid;
-use ipld_blockstore::BlockStore;
-use ipld_hamt::{Error as HamtError, Hamt};
 use num_derive::FromPrimitive;
 use num_traits::FromPrimitive;
 use runtime::{ActorCode, Runtime};
-use serde::{Deserialize, Serialize};
-
-/// InitActorState is reponsible for creating
-// TODO implement actual serialize and deserialize to match
-#[derive(Serialize, Deserialize)]
-pub struct InitActorState {
-    address_map: Cid,
-    next_id: ActorID,
-}
-
-impl InitActorState {
-    pub fn new(address_map: Cid) -> Self {
-        Self {
-            address_map,
-            next_id: FIRST_NON_SINGLETON_ADDR,
-        }
-    }
-    /// Assigns next available ID and incremenets the next_id value from state
-    pub fn map_address_to_new_id<BS: BlockStore>(
-        &mut self,
-        store: &BS,
-        addr: &Address,
-    ) -> Result<Address, HamtError> {
-        let id = self.next_id;
-        self.next_id += 1;
-
-        let mut map: Hamt<String, _> = Hamt::load_with_bit_width(&self.address_map, store, 5)?;
-        map.set(String::from_utf8_lossy(&addr.to_bytes()).to_string(), id)?;
-        self.address_map = map.flush()?;
-
-        Ok(Address::new_id(id.0).expect("Id Address should be created without Error"))
-    }
-
-    /// Resolve address
-    pub fn resolve_address<BS: BlockStore>(
-        &self,
-        _store: &BS,
-        _addr: &Address,
-    ) -> Result<Address, String> {
-        // TODO implement address resolution
-        todo!()
-    }
-}
+use vm::{CodeID, ExitCode, MethodNum, Serialized, METHOD_CONSTRUCTOR, METHOD_PLACEHOLDER};
 
 #[derive(FromPrimitive)]
 pub enum InitMethod {
@@ -71,8 +24,8 @@ impl InitMethod {
     }
 }
 
-pub struct InitActorCode;
-impl InitActorCode {
+pub struct InitActor;
+impl InitActor {
     fn constructor<RT: Runtime>(_rt: &RT) {
         // Acquire state
         // Update actor substate
@@ -86,7 +39,7 @@ impl InitActorCode {
     }
 }
 
-impl ActorCode for InitActorCode {
+impl ActorCode for InitActor {
     fn invoke_method<RT: Runtime>(&self, rt: &RT, method: MethodNum, params: &Serialized) {
         // Create mutable copy of params for usage in functions
         let params: &mut Serialized = &mut params.clone();
