+++ vm/actor/src/builtin/miner/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+mod state;
+
+pub use self::state::{MinerActorState, MinerInfo};
+
+pub struct MinerActor;
