+++ vm/actor/src/builtin/miner/state.rs
@@ -7,10 +7,8 @@ use encoding::de;
 use libp2p::PeerId;
 use serde::Deserialize;
 
-pub struct StorageMinerActor {}
-
 /// Container representing storage miner actor state
-pub struct StorageMinerActorState {
+pub struct MinerActorState {
     // TODO add proving_set, post_state
     /// The height at which this miner was slashed at.
     pub slashed_at: u64,
@@ -18,6 +16,20 @@ pub struct StorageMinerActorState {
     pub sectors: Cid,
     /// Contains static info about this miner
     pub info: Cid,
+}
+
+impl<'de> de::Deserialize<'de> for MinerActorState {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: de::Deserializer<'de>,
+    {
+        let (slashed_at, sectors, info) = Deserialize::deserialize(deserializer)?;
+        Ok(Self {
+            slashed_at,
+            sectors,
+            info,
+        })
+    }
 }
 
 /// Static information about miner
@@ -41,20 +53,6 @@ impl MinerInfo {
     /// Returns a reference of the amount of space in each sector committed to the network by this miner
     pub fn sector_size(&self) -> &u64 {
         &self.sector_size
-    }
-}
-
-impl<'de> de::Deserialize<'de> for StorageMinerActorState {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        let (slashed_at, sectors, info) = Deserialize::deserialize(deserializer)?;
-        Ok(Self {
-            slashed_at,
-            sectors,
-            info,
-        })
     }
 }
