+++ vm/actor/src/builtin/power/state.rs
@@ -1,0 +1,10 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use num_bigint::BigUint;
+
+/// State of storage power actor
+pub struct PowerActorState {
+    // TODO add power tables on finished spec
+    _total_storage: BigUint,
+}
